Revert some unneeded code of previous commit
There is no need to check all port numbers again when decoding the version
number. Only port == 10 must be treated specially (which did not have the
version number), all other ports just skip the version byte. Also, an error for
an unsupported port is not needed there, since that was checked already
previously.