Use port numbers more when decoding
Previously, some values were decoded when the packet was long enough, rather
than based on the port number. This happened to work with the current
combination of packets being received, but could break with future changes.
Checking for the port number also makes it more explicit what kind of packets
are expected on what ports.

Decoding the battery level still happens based on the length, since the
presence of the battery level is not reflected in the port number.